backend/requirements.txt
aniso8601==8.0.0
astroid==2.4.2
attrs==20.3.0
autopep8==1.5.4
beautifulsoup4==4.9.1
branca==0.4.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
folium==0.11.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.4.0
iniconfig==1.1.1
install==1.3.4
isort==5.5.3
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
lxml==4.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.1
packaging==20.8
pandas==1.1.1
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.6.0
pylint==2.6.0
PyMySQL==0.10.1
pyparsing==2.4.7
pytest==6.2.1
pytest-flask==1.1.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
